fix: typescript errors
